+++ migrations/versions/20221209_180934_all_tables_or_some_shit.py
@@ -37,7 +37,7 @@ def upgrade():
     op.create_table('carts',
     sa.Column('id', sa.Integer(), nullable=False),
     sa.Column('user_id', sa.Integer(), nullable=False),
-    sa.Column('total', sa.Integer(), nullable=False),
+    sa.Column('total', sa.NUMERIC(10, 0), nullable=False),
     sa.Column('purchased', sa.Boolean(), nullable=True),
     sa.ForeignKeyConstraint(['user_id'], ['users.id'], ),
     sa.PrimaryKeyConstraint('id')
@@ -49,7 +49,7 @@ def upgrade():
     sa.Column('detailed_description', sa.String(length=5000), nullable=True),
     sa.Column('category_id', sa.Integer(), nullable=False),
     sa.Column('owner_id', sa.Integer(), nullable=False),
-    sa.Column('price', sa.Integer(), nullable=False),
+    sa.Column('price', sa.NUMERIC(10, 0), nullable=False),
     sa.Column('preview_img_id', sa.Integer(), nullable=False),
     sa.ForeignKeyConstraint(['category_id'], ['categories.id'], ),
     sa.ForeignKeyConstraint(['owner_id'], ['users.id'], ),
